Converted the library to generate json output
Basically modified the jar file to move away from generating generally
inflexible html to a json storage medium so that a html / javascript
renderer can customize page output much more easily.

The output is now stored in data.js as var javadoc = { …. }, and the
render.js iterates through the data set and embeds it into div#content
in index.html.

Currently all that is supported is display all javadoc in a singular
page, but the data set will now allow for easier searching, dynamic
rendering.  Speed appears to be fine on a medium-large sized codebase